useCommands hook - disabled property
